REFACTOR: discussion 관련 함수 분리
